trabalho01/seguranca/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
idna==3.3
Naked==0.1.31
pycryptodome==3.14.1
python-decouple==3.6
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
